files/website/requirements.txt
Django==4.1.10
django-libsass==0.9
django-storages[google]==1.14
django-sekizai==4.1.0
httpx==0.24.1
python-dotenv==1.0.0
pytz==2023.3
psycopg2-binary==2.9.7
uvicorn==0.23.2
